moving scripts around
