+++ Assets/Scripts/Drops.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a53226af3a80d2c44b7854289458fa3e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
