+++ Assets/Scripts/Drops/Drop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 833242e46493fc043acf1a9a891d344a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
