+++ Assets/Scripts/Drops/LimbDrop.cs
@@ -19,32 +19,18 @@ public class LimbDrop : MonoBehaviour
     public float LimbHealth { get { return limbHealth; } }
 
 
-    private void OnEnable()
-    {
-        //if (limbHealth == 0)
-        //{
-        //    limbHealth = 100f;
-        //}
-        //Debug.Log("limbHealth: " + limbHealth);
+    // private void OnEnable()
+    // {
+    //     //if (limbHealth == 0)
+    //     //{
+    //     //    limbHealth = 100f;
+    //     //}
+    //     //Debug.Log("limbHealth: " + limbHealth);
 
-    }
-
-    private void Start()
-    {
-        Vector3 forceDirection = PointOnXZCircle(transform.position, 1f, Random.Range(0,360)).normalized * 20f + Vector3.up * 100f;
-        GetComponent<Rigidbody>().AddForce(forceDirection, ForceMode.Impulse);
-    }
+    // }
 
     public void OverwriteLimbHealth(float newValue)
     {
         limbHealth = newValue;
     }
-
-    private Vector3 PointOnXZCircle(Vector3 center, float radius, float angle)
-    {
-        float a = angle * Mathf.PI / 180f;
-        return center + new Vector3(Mathf.Sin(a), 0, Mathf.Cos(a)) * radius;
-    }
-
-    
 }
